+++ network/data/src/test/java/me/proton/core/network/data/ApiManagerTests.kt
@@ -142,6 +142,7 @@ internal class ApiManagerTests {
                 mockk(),
                 scope,
                 cache = { null },
+                apiConnectionListener = null
             )
 
         coEvery { dohService.getAlternativeBaseUrls(any()) } returns listOf(proxy1url)